Fixed one of two major stupid errors.  First, there is no need to separately compile implementation.c, since it is **included** into signal.pyx by the line "cdef extern from "implementation.c".  This commit fixes that.
The second issue is that, currently, both the pari_gen extension and the
cysignals extension are being linked agains the pari static library.  That
means that the global symbols, such as PARI_SIGINT_pending and
PARI_SIGINT_block, which are used by cysignals are distinct from the ones
used in calls to the pari library from cypari.pari.  In other words,
cysignals is not actually working at the moment.  Either cysignals
should be embedded in pari_gen, or these global symbols should somehow
be exported by pari_gen.